third_party/blink/web_tests/fast/events/constructors/midi-connection-event-constructor-expected.txt

This tests the constructor for the MIDIConnectionEvent DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new MIDIConnectionEvent('eventType').bubbles is false
PASS new MIDIConnectionEvent('eventType').cancelable is false
PASS new MIDIConnectionEvent('eventType').port is null
PASS new MIDIConnectionEvent('eventType', { bubbles: false }).bubbles is false
PASS new MIDIConnectionEvent('eventType', { bubbles: true }).bubbles is true
PASS new MIDIConnectionEvent('eventType', { cancelable: false }).cancelable is false
PASS new MIDIConnectionEvent('eventType', { cancelable: true }).cancelable is true
PASS new MIDIConnectionEvent('eventType', { bubbles: true, cancelable: true }).bubbles is true
PASS new MIDIConnectionEvent('eventType', { bubbles: true, cancelable: true }).cancelable is true
PASS successfullyParsed is true

TEST COMPLETE
